requirements.txt
reflex==0.7.14
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
bcrypt==4.0.1
plotly==5.17.0
docx2txt==0.8